[all] Issue 104 fixed (inputs instead of sticks for BL activation). Should spare some flash on stock, but I didn't check [Taranis] No scrollbar on failsafe screen (and some arrangements to keep consistency with other menus) [Taranis] PPM_UNIT=PREC0 / US taken into account in channels view

+++ src/opentx.h
@@ -729,7 +729,7 @@ void watchdogSetTimeout(uint32_t timeout);
 struct t_inactivity
 {
   uint16_t counter;
-  uint16_t sum;
+  uint8_t  sum;
 };
 extern struct t_inactivity inactivity;
